Step4 : Merge GlobalredirectSubscriber, fix unit test to pass, change in schema because of core changes.
